"修复逻辑错误"

+++ src/main/java/com/xd/hotel/service/StatisticService.java
@@ -6,7 +6,7 @@ import java.time.LocalDateTime;
  * Created by jiaqi on 2019/6/14 2:10 PM
  */
 public interface StatisticService {
-    Long getIncome(LocalDateTime fromDate);
+    Integer getIncome(LocalDateTime fromDate);
 
     Integer getCartCount(String identityNumber);
 }
